chore: add issue templates and centralize fixtures (ref #71)
