+++ .github/ISSUE_TEMPLATE/feature_request.md
@@ -1,0 +1,17 @@
+---
+name: Feature Request
+about: Suggest an idea for this project
+labels: enhancement
+---
+
+## Problem
+[Is your feature request related to a problem? Please describe.]
+
+## Proposed Solution
+[Describe the solution you'd like]
+
+## Alternatives Considered
+[Describe alternatives you've considered]
+
+## Additional Context
+[Add any other context or screenshots about the feature request here.]
